+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Bundle optimizations
+- Styles for notifications
+
 ## [1.5.0] - 2021-11-27
 
 ### Added